refactor list items to own component
