+++ components/ItemCard/index.js
@@ -1,0 +1,3 @@
+import { ItemCard } from "./ItemCard";
+
+export default ItemCard;
